requirements.txt
torch
transformers
psycopg2
numpy
streamlit
fastapi
uvicorn
python-multipart
python-dotenv